progress with character topics
