+++ src/main/python/plotlyst/core/client.py
@@ -47,7 +47,7 @@ from plotlyst.core.domain import Novel, Character, Scene, Chapter, SceneStage, \
     default_character_networks, ScenePurposeType, StoryElement, SceneOutcome, ChapterType, SceneStructureItem, \
     DocumentProgress, ReaderQuestion, SceneReaderQuestion, ImageRef, SceneReaderInformation, \
     CharacterProfileSectionReference, CharacterMultiAttribute, default_character_profile, CharacterPersonality, \
-    StrengthWeaknessAttribute, PremiseBuilder, SceneFunctions, Location, default_locations
+    StrengthWeaknessAttribute, PremiseBuilder, SceneFunctions, Location, default_locations, TopicElement
 from plotlyst.core.template import Role, exclude_if_empty, exclude_if_black, exclude_if_false
 from plotlyst.env import app_env
 
@@ -137,7 +137,7 @@ class CharacterInfo:
     document: Optional[Document] = None
     journals: List[Document] = field(default_factory=list)
     prefs: CharacterPreferences = field(default_factory=CharacterPreferences)
-    topics: List[TemplateValue] = field(default_factory=list)
+    topics: List[TopicElement] = field(default_factory=list)
     big_five: Dict[str, List[int]] = field(default_factory=default_big_five_values)
     profile: List[CharacterProfileSectionReference] = field(default_factory=default_character_profile)
     summary: str = field(default='', metadata=config(exclude=exclude_if_empty))
